activated dependencies are input to read only operations so don't dup.
We pass activated to sort_dependencies and DependencyState.new, both operations
that probably shouldn't be modifying any of the dependency graphs.

Note: activated contains DependencyGraph objects and calling .dup is not a fast
operation because it has to copy the graph and rebuild the vertices.

For a large Rails application that runs Molinillo::Resolver::Resolution#attempt_to_activate
477 times when running bundle or require 'bundler/setup', this saves roughly 10% time.