Consider dependency possibilities in groups
Dependencies usually don't change their sub-dependency requirements when they
update. As such, we can treat several versions of the same dependency as
functionally equivalent when resolving the dependency tree.

To exploit the above, we introduce a PossibilitySet class, which holds an array
of dependency possibilities, all of which have the same sub-dependencies.
Vertices are pinned to a PossibilitySet, rather than a specific version, during
resolution, with the set filtered as required throughout the process. A concrete
version of each dependency is then chosen once resolution as concluded.