Retain all previous unwind options
Previously, we kept a hash of previous conflicts in `DependencyState#conflicts`,
and considered unwinds from those previous considering when determining which
state to unwind to.

Unfortunately, there were a couple of problems with that approach:
- A maximum of one conflict per dependency was stored. When a conflict was
  "resolved" and a new version of that dependency was activated, the previous
  conflict was deleted. Doing so deleted unwind options that may have been used
  in the future.
- The `DependencyState#conflicts` method was also used when constructing error
  messages, so our ability to tweak it's behaviour was limited.
- Full requirement trees for the conflicting requirements, and the associated
  unwind options, weren't stored, and could not be recreated at a later date
  (since the DependencyGraph would have changed since the time when the conflict
  was created)

This PR introduces a new store on each `DependencyState` which contains details
of all previously unused unwinds. Those old, previously unused unwinds are then
considered when deciding whether to unwind from a new conflict. In theory, an
old unwind may allow us to avoid the conflict we've just encountered, if it
would affect the conflict's requirement trees in any way.