Allow fully customizing molinillo's error building
Currently bundler needs to passed 7 different procs as options, and
the logic is split between those procs and molinillo's internals which
makes it really hard to follow.

In addition to that, the above is not even enough for our needs. We have
some "metadata dependencies" (the running Ruby and RubyGems version),
which are not set in the dependency manifest (Gemfile), so we need to
customize the message so that it does not read "In Gemfile:". However,
that part is not currently customizable.

This commit adds a new proc that allows customizing building the full
error message.